requirements.txt
gradio==5.8.0
librosa==0.9.0
matplotlib==3.9.2
numpy==1.26.4
tensorflow==2.17.0
google-generativeai==0.8.3